// FrontEnd/src/app/services/api-backend/model/usuario.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AplicacionesMedicamento } from './aplicacionesMedicamento';
import { RolesUsuario } from './rolesUsuario';
import { Cita } from './cita';
import { MedicamentosCaso } from './medicamentosCaso';
import { TiposUsuario } from './tiposUsuario';
import { Venta } from './venta';
import { Receta } from './receta';
import { DiagnosticosCaso } from './diagnosticosCaso';
import { ExamenesCaso } from './examenesCaso';


export interface Usuario { 
    idUsuario?: number;
    nombres?: string | null;
    apellidos?: string | null;
    fechaNacimiento?: string;
    estado?: string | null;
    username?: string | null;
    password?: string | null;
    idTipoUsuario?: number;
    idTipoUsuarioNavigation?: TiposUsuario;
    aplicacionesMedicamentos?: Array<AplicacionesMedicamento> | null;
    cita?: Array<Cita> | null;
    diagnosticosCasos?: Array<DiagnosticosCaso> | null;
    examenesCasos?: Array<ExamenesCaso> | null;
    medicamentosCasos?: Array<MedicamentosCaso> | null;
    receta?: Array<Receta> | null;
    rolesUsuarios?: Array<RolesUsuario> | null;
    venta?: Array<Venta> | null;
}